refactor: Give capture the same per-generation adapter as restore
export_bulk_config keeps the export envelope and the per-device loop; how one generation's component configs are captured (Gen2 GetConfig/GetCode/Schedule.List vs the Gen1 mapped-config copy plus raw /settings fetch) moves unchanged into ComponentCaptureStrategy implementations under use_cases/capture_strategies.

LEGACY_SETTINGS_KEY moves to the device_backup entity: it names part of the snapshot format, which capture and restore now both consume from the same home.
