+++ packages/core/src/core/use_cases/capture_strategies/gen1.py
@@ -1,0 +1,46 @@
+"""Capture strategy for Gen1 (legacy HTTP) devices."""
+
+from typing import Any
+
+from core.domain.entities.device_backup import LEGACY_SETTINGS_KEY
+from core.domain.entities.device_status import DeviceStatus
+from core.gateways.device import DeviceGateway
+
+
+class Gen1CaptureStrategy:
+    """Capture Gen1 component configs plus the raw ``/settings``.
+
+    Gen1 has no /rpc: GetConfig and Schedule.List 404, so the mapped configs
+    already on the ``DeviceStatus`` are captured instead. The
+    ``legacy_settings`` entry is the source of truth a Gen1 restore replays;
+    it is omitted when the raw fetch fails, and the mapped configs alone still
+    make the backup valid.
+    """
+
+    def __init__(self, device_gateway: DeviceGateway):
+        self._device_gateway = device_gateway
+
+    async def capture_components(
+        self, device_ip: str, status: DeviceStatus, component_types: list[str]
+    ) -> dict[str, Any]:
+        components: dict[str, Any] = {}
+
+        for component in status.components:
+            if component.component_type in component_types:
+                components[component.key] = {
+                    "type": component.component_type,
+                    "success": True,
+                    "config": component.config,
+                    "error": None,
+                }
+
+        legacy_settings = await self._device_gateway.get_legacy_settings(device_ip)
+        if legacy_settings is not None:
+            components[LEGACY_SETTINGS_KEY] = {
+                "type": LEGACY_SETTINGS_KEY,
+                "success": True,
+                "config": legacy_settings,
+                "error": None,
+            }
+
+        return components
